alterado tipo do atributo weightDone (int para double)

+++ src/main/java/com/mixsz/workouttracker/dto/response/WorkoutLogExerciseResponseDTO.java
@@ -2,6 +2,6 @@ package com.mixsz.workouttracker.dto.response;
 
 import java.util.UUID;
 
-public record WorkoutLogExerciseResponseDTO(UUID id, ExerciseResponseDTO exercise, int weightDone, int setsDone,
+public record WorkoutLogExerciseResponseDTO(UUID id, ExerciseResponseDTO exercise, double weightDone, int setsDone,
                                             int repsDone, int position) {
 }
